Add unit test for CloseTimerTask,ReconnectTimerTask,HeaderExchangeServer

+++ dubbo-remoting/dubbo-remoting-api/src/main/java/org/apache/dubbo/remoting/exchange/support/header/HeaderExchangeServer.java
@@ -103,7 +103,7 @@ public class HeaderExchangeServer implements ExchangeServer {
     public void close(final int timeout) {
         startClose();
         if (timeout > 0) {
-            final long max = (long) timeout;
+            final long max = timeout;
             final long start = System.currentTimeMillis();
             if (getUrl().getParameter(Constants.CHANNEL_SEND_READONLYEVENT_KEY, true)) {
                 sendChannelReadOnlyEvent();
